(PC-1537) refactor name to make diff beetwen vars, add tests, split files
