+++ src/components/pages/Offer/EventOccurrenceAndStockItem/EditAndDeleteActions.jsx
@@ -8,7 +8,7 @@ import { requestData } from 'redux-saga-data'
 
 import DeleteDialog from './DeleteDialog'
 
-class Actions extends Component {
+class EditAndDeleteActions extends Component {
   constructor(props) {
     super(props)
     this.state = {
@@ -44,6 +44,12 @@ class Actions extends Component {
   render() {
     const { offer, isStockOnly, stockPatch, eventOccurrencePatch } = this.props
 
+    const editAndDeleteNavUrl = `/offres/${get(offer, 'id')}?gestion&${
+      isStockOnly
+        ? `stock=${get(stockPatch, 'id')}`
+        : `date=${get(eventOccurrencePatch, 'id')}`
+    }`
+
     // Delete dialog
     if (this.state.isDeleting) {
       return (
@@ -64,11 +70,7 @@ class Actions extends Component {
       <Fragment>
         <td>
           <NavLink
-            to={`/offres/${get(offer, 'id')}?gestion&${
-              isStockOnly
-                ? `stock=${get(stockPatch, 'id')}`
-                : `date=${get(eventOccurrencePatch, 'id')}`
-            }`}
+            to={editAndDeleteNavUrl}
             className="button is-small is-secondary edit-stock">
             <span className="icon">
               <Icon svg="ico-pen-r" />
@@ -92,4 +94,4 @@ class Actions extends Component {
   }
 }
 
-export default connect()(Actions)
+export default connect()(EditAndDeleteActions)
