+++ src/components/pages/Offer/EventOccurrenceAndStockItem/tests/mockedState.js
@@ -1051,216 +1051,7 @@ const mockedState = {
     isActive: false,
   },
   notification: null,
-  tracker: {
-    'users/current': '2019-03-07T13:02:50.482Z',
-    types: '2019-03-07T10:44:31.201Z',
-    'offers?page=1&orderBy=offer.id+desc': '2019-02-25T13:40:30.827Z',
-    'offerers?validated=false': '2019-03-06T12:49:58.924Z',
-    'offerers?page=1&': '2019-02-27T09:37:55.642Z',
-    'offerers?page=2&': '2019-02-12T10:33:56.969Z',
-    'offerers?page=3&': '2019-02-12T10:33:59.867Z',
-    providers: '2019-03-07T13:02:50.907Z',
-    'offers/AJMQ': '2019-01-11T13:19:18.740Z',
-    'users/signout': '2019-02-15T08:29:26.862Z',
-    'userOfferers/APJQ': '2019-02-01T13:59:11.142Z',
-    'offerers/APJQ': '2019-02-01T13:59:11.170Z',
-    'userOfferers/APEQ': '2019-02-01T14:52:52.945Z',
-    'offerers/APEQ': '2019-02-01T14:52:52.949Z',
-    'venueProviders?venueId=AQYQ': '2019-02-01T14:52:52.927Z',
-    'userOfferers/APYA': '2019-02-05T11:19:52.028Z',
-    'offerers/APYA': '2019-02-05T11:19:52.030Z',
-    'offerers/APWA': '2019-02-05T11:17:40.795Z',
-    'userOfferers/APWA': '2019-02-05T11:17:40.744Z',
-    'venueProviders?venueId=ARRA': '2019-02-05T11:17:32.674Z',
-    'userOfferers/APYQ': '2019-02-05T11:29:00.492Z',
-    'offerers/APYQ': '2019-02-05T11:29:00.508Z',
-    'userOfferers/APXQ': '2019-02-05T12:50:53.885Z',
-    'offerers/APXQ': '2019-02-05T12:50:53.954Z',
-    'venueProviders?venueId=ARTQ': '2019-02-05T12:35:00.739Z',
-    'offers?page=1&offererId=APSQ&orderBy=offer.id+desc':
-      '2019-02-05T11:47:37.058Z',
-    'userOfferers/APVQ': '2019-02-05T12:02:05.376Z',
-    'offerers/APVQ': '2019-02-05T12:02:05.405Z',
-    'venueProviders?venueId=ARQA': '2019-02-05T12:00:05.805Z',
-    'venueProviders?venueId=ARWQ': '2019-02-05T12:54:31.014Z',
-    'venueProviders?venueId=AR2Q': '2019-02-05T13:46:45.748Z',
-    'userOfferers/AP7Q': '2019-02-11T09:15:27.624Z',
-    'offerers/AP7Q': '2019-02-11T09:15:27.628Z',
-    'venueProviders?venueId=ASEQ': '2019-02-11T09:15:26.608Z',
-    'userOfferers/BQ': '2019-02-11T10:18:10.790Z',
-    'offerers/BQ': '2019-02-11T10:18:10.807Z',
-    'venueProviders?venueId=CM': '2019-02-11T10:18:10.744Z',
-    'offers/AKKQ': '2019-02-12T10:32:07.243Z',
-    'userOfferers/NU': '2019-02-12T10:33:41.549Z',
-    'offerers/NU': '2019-02-12T10:39:25.870Z',
-    'offers/AKLA': '2019-02-12T10:45:44.226Z',
-    'offerers/N9': '2019-02-12T10:39:02.346Z',
-    'offers/AKLQ': '2019-02-12T10:38:31.791Z',
-    'offerers/MM': '2019-02-12T10:38:54.772Z',
-    'offers/AKMA': '2019-02-12T10:44:29.853Z',
-    'offers/AUXQ': '2019-02-13T14:30:34.924Z',
-    'offers?page=2&orderBy=offer.id+desc': '2019-02-13T14:34:52.412Z',
-    'offers?page=3&orderBy=offer.id+desc': '2019-02-13T14:34:59.006Z',
-    'offers/AUNQ': '2019-02-13T14:34:36.942Z',
-    'offers?page=4&orderBy=offer.id+desc': '2019-02-13T14:35:04.356Z',
-    'offers?page=5&orderBy=offer.id+desc': '2019-02-13T14:35:09.840Z',
-    'offers?page=6&orderBy=offer.id+desc': '2019-02-13T14:35:13.784Z',
-    'offers?page=7&orderBy=offer.id+desc': '2019-02-13T14:35:18.026Z',
-    'offers?page=8&orderBy=offer.id+desc': '2019-02-13T14:35:22.709Z',
-    'offers/ATWQ': '2019-02-13T14:40:27.725Z',
-    'offers/AWNA': '2019-02-13T15:01:32.939Z',
-    'offerers/AH9Q': '2019-02-13T15:01:49.545Z',
-    'offers/AWNQ': '2019-02-13T15:06:11.468Z',
-    'userOfferers/ALHQ': '2019-02-14T12:28:25.474Z',
-    'offerers/ALHQ': '2019-02-14T12:28:25.493Z',
-    'offers/BDEA': '2019-02-14T12:29:25.335Z',
-    'offers/ME': '2019-02-22T13:34:55.863Z',
-    "offers?page=1&keywords=DANSONS JUSQU'EN 2030 (TOTOT)&orderBy=offer.id+desc":
-      '2019-02-25T13:01:27.867Z',
-    "offers?page=1&keywords=DANSONS JUSQU'EN 2030 &orderBy=offer.id+desc":
-      '2019-02-22T13:34:25.622Z',
-    "offers?page=1&keywords=Dansons jusqu'en 2030  (TOTOT)&orderBy=offer.id+desc":
-      '2019-02-22T13:34:34.259Z',
-    'offers?page=1&keywords=Avez-vous déjà vu ?&orderBy=offer.id+desc':
-      '2019-02-22T13:35:09.458Z',
-    'offers?page=1&keywords=DANSONS%20JUSQU%27EN%202030%20(TOTOT)&orderBy=offer.id+desc':
-      '2019-02-25T10:22:39.301Z',
-    'offers?page=1&orderBy=offer.id%20desc': '2019-02-25T14:04:13.335Z',
-    'offers?page=2&orderBy=offer.id%20desc': '2019-02-25T13:53:19.374Z',
-    'offers?page=1&?orderBy=offer.id%20desc': '2019-02-25T14:30:31.170Z',
-    'offers/NE': '2019-02-27T08:28:35.841Z',
-    'offers?page=1&?orderBy=offer.id+desc': '2019-02-25T14:05:25.086Z',
-    'offers?page=1&?orderBy=offer.id%20desc&search=titi':
-      '2019-02-25T14:09:33.699Z',
-    'offers?page=1&?orderBy=offer.id%20desc&search=Sun%20aux%20lentilles':
-      '2019-02-25T14:06:03.327Z',
-    'offers?page=1&?page=1&orderBy=offer.id%20desc': '2019-02-25T14:21:40.068Z',
-    'offers?page=2&?orderBy=offer.id%20desc&page=1&search=uyrytdlyrd':
-      '2019-02-25T14:06:17.301Z',
-    'offers?page=2&?orderBy=offer.id%20desc&page=1&search=titiYTERTSRU':
-      '2019-02-25T14:09:39.257Z',
-    'offers?page=2&?page=1&orderBy=offer.id%20desc': '2019-02-25T14:21:43.552Z',
-    'offers?page=3&?page=2&orderBy=offer.id%20desc': '2019-02-25T14:17:58.405Z',
-    'offers?page=4&?page=3&orderBy=offer.id%20desc': '2019-02-25T14:18:00.416Z',
-    'offers?page=5&?page=4&orderBy=offer.id%20desc': '2019-02-25T14:18:02.742Z',
-    'offers?page=6&?orderBy=offer.id%20desc&search=oc%C3%A9an':
-      '2019-02-25T14:18:07.751Z',
-    'offers?page=1&?orderBy=offer.id%20desc&search=oc%C3%A9an':
-      '2019-02-25T14:19:26.314Z',
-    'offers?page=1&?keywords=titi&orderBy=offer.id%20desc':
-      '2019-02-25T14:19:29.568Z',
-    'offers?page=1&?keywords=oc%C3%A9an&orderBy=offer.id%20desc':
-      '2019-02-25T14:19:43.041Z',
-    'offers?page=1&?keywords=rtyiryioc%C3%A9an&orderBy=offer.id%20desc':
-      '2019-02-25T14:19:58.832Z',
-    'offers?page=1&?keywords=ttoispmat&orderBy=offer.id%20desc':
-      '2019-02-25T14:20:12.709Z',
-    'offers?page=1&?keywords=aloiyufygcx&orderBy=offer.id%20desc':
-      '2019-02-25T14:21:40.177Z',
-    'offers?page=1&?page=2&orderBy=offer.id%20desc': '2019-02-25T14:22:32.240Z',
-    'offers?page=2&?orderBy=offer.id%20desc': '2019-02-25T14:22:32.369Z',
-    'offers?': '2019-03-07T10:44:32.156Z',
-    'offers?orderBy=offer.id%20desc': '2019-03-07T10:44:32.298Z',
-    'offers?keywords=titi&orderBy=offer.id%20desc': '2019-02-26T15:17:20.354Z',
-    'offers?keywords=sun%20aux%20lentilles&orderBy=offer.id%20desc':
-      '2019-03-06T13:14:20.482Z',
-    'offers?orderBy=offer.id%20desc&page=1': '2019-02-25T15:38:20.351Z',
-    'offers?orderBy=offer.id%20desc&page=2': '2019-03-06T12:56:42.547Z',
-    'offers?orderBy=offer.id%20desc&page=3': '2019-03-06T12:56:46.611Z',
-    'offers?orderBy=offer.id%20desc&page=4': '2019-02-27T09:37:48.822Z',
-    'offers?orderBy=offer.id%20desc&page=5': '2019-02-27T09:37:51.179Z',
-    'offers?keywords=oc%C3%A9an&orderBy=offer.id%20desc':
-      '2019-03-06T13:44:05.267Z',
-    'offers?keywords=ytritss&orderBy=offer.id%20desc':
-      '2019-02-25T14:39:15.709Z',
-    'offers?keywords=lentilles&orderBy=offer.id%20desc':
-      '2019-02-25T15:11:09.268Z',
-    'offers?keywords=gjvl&orderBy=offer.id%20desc': '2019-02-25T15:10:32.944Z',
-    'offers?keywords=oc%C3%A9an&orderBy=offer.id%20desc&page=3':
-      '2019-03-06T13:42:48.711Z',
-    'offers?keywords=oc%C3%A9an&orderBy=offer.id%20desc&page=4':
-      '2019-02-26T09:25:43.818Z',
-    'offers?keywords=sun%20aux%20lentilles&orderBy=offer.id%20desc&page=2':
-      '2019-02-26T15:17:27.882Z',
-    'offers?keywords=sun%20aux%20lentilles&orderBy=offer.id%20desc&page=3':
-      '2019-03-06T13:14:21.662Z',
-    'offers?keywords=utuu&orderBy=offer.id%20desc': '2019-02-26T09:36:03.970Z',
-    'offers?keywords=sun&orderBy=offer.id%20desc': '2019-02-26T09:40:33.815Z',
-    'offers?orderBy=offer.id%20desc&page=6': '2019-02-26T13:59:12.217Z',
-    'offers?orderBy=offer.id%20desc&page=7': '2019-02-26T13:56:50.057Z',
-    'offers?keywords=oc%C3%A9an&orderBy=offer.id%20desc&page=2':
-      '2019-03-06T13:42:46.993Z',
-    'offers/M4': '2019-02-26T12:54:42.960Z',
-    'offers?keywords=sun%20aux%20lentilles&orderBy=offer.id%20desc&page=4':
-      '2019-03-06T13:14:21.597Z',
-    'offers?keywords=sun%20aux%20lentilles': '2019-02-26T09:51:20.340Z',
-    'offers?keywords=sun%20aux%20lentilles&page=2': '2019-02-26T09:51:22.141Z',
-    'offers?keywords=sun%20aux%20lentilles&page=3': '2019-02-26T09:51:22.926Z',
-    'offers?page=2': '2019-02-26T12:25:00.966Z',
-    'offers?structure=A9': '2019-02-26T13:51:54.866Z',
-    'offers?orderBy=offer.id%20desc&structure=A9': '2019-02-26T13:51:54.964Z',
-    'offers?structure=BU': '2019-02-27T09:07:57.111Z',
-    'offers?orderBy=offer.id%20desc&structure=BU': '2019-02-27T09:11:07.864Z',
-    'offers?orderBy=offer.id%20desc&page=2&structure=BU':
-      '2019-02-27T08:33:19.638Z',
-    'offers?orderBy=offer.id%20desc&page=3&structure=BU':
-      '2019-02-27T08:33:19.758Z',
-    'offers?orderBy=offer.id%20desc&page=4&structure=BU':
-      '2019-02-27T08:29:01.895Z',
-    'offers?orderBy=offer.id%20desc&page=5&structure=BU':
-      '2019-02-27T08:29:02.720Z',
-    'offers?orderBy=offer.id%20desc&page=6&structure=BU':
-      '2019-02-27T08:29:03.589Z',
-    'offers?orderBy=offer.id%20desc&page=7&structure=BU':
-      '2019-02-27T08:29:04.368Z',
-    'offers?orderBy=offer.id%20desc&page=8&structure=BU':
-      '2019-02-27T08:29:05.332Z',
-    'offers?orderBy=offer.id%20desc&page=9&structure=BU':
-      '2019-02-27T08:29:06.247Z',
-    'offers?orderBy=offer.id%20desc&page=10&structure=BU':
-      '2019-02-27T08:29:07.256Z',
-    'offers?orderBy=offer.id%20desc&page=11&structure=BU':
-      '2019-02-27T08:29:07.990Z',
-    'offers?orderBy=offer.id%20desc&page=12&structure=BU':
-      '2019-02-27T08:29:08.546Z',
-    'offers?orderBy=offer.id%20desc&page=13&structure=BU':
-      '2019-02-27T08:29:09.165Z',
-    'offers?keywords=oc%C3%A9an&orderBy=offer.id%20desc&structure=BU':
-      '2019-02-27T09:10:01.553Z',
-    'offers?keywords=oc%C3%A9an&orderBy=offer.id%20desc&page=3&structure=BU':
-      '2019-02-27T09:10:02.134Z',
-    'offers?keywords=oc%C3%A9an&orderBy=offer.id%20desc&page=4&structure=BU':
-      '2019-02-27T09:10:02.182Z',
-    'offers?structure=AY': '2019-02-27T09:11:31.926Z',
-    'offers?orderBy=offer.id%20desc&structure=AY': '2019-02-27T09:11:32.014Z',
-    'offers?structure=BQ': '2019-02-27T09:12:19.701Z',
-    'offers?orderBy=offer.id%20desc&structure=BQ': '2019-02-27T09:12:19.807Z',
-    'offers?structure=AM': '2019-02-27T09:12:30.259Z',
-    'offers?orderBy=offer.id%20desc&structure=AM': '2019-02-27T09:22:11.271Z',
-    'offers?offererId=AM&orderBy=offer.id%20desc': '2019-02-27T09:36:59.371Z',
-    'offers?offererId=AM': '2019-02-27T09:30:12.580Z',
-    'offers?offererId=AM&orderBy=offer.id%20desc&page=2':
-      '2019-02-27T09:29:19.054Z',
-    'offers?offererId=AM&orderBy=offer.id%20desc&page=3':
-      '2019-02-27T09:29:19.499Z',
-    'offers?keywords=oc%C3%A9an&offererId=AM&orderBy=offer.id%20desc':
-      '2019-02-27T09:36:58.953Z',
-    'offers/G4': '2019-02-27T09:37:22.089Z',
-    'offers?offererId=BU': '2019-02-27T09:48:24.205Z',
-    'offers?offererId=BU&orderBy=offer.id%20desc': '2019-02-27T11:14:47.988Z',
-    'offers?offererId=BU&orderBy=offer.id%20desc&page=2':
-      '2019-02-27T10:10:35.954Z',
-    'offers?offererId=BU&orderBy=offer.id%20desc&page=3':
-      '2019-02-27T10:10:36.315Z',
-    'offerers?': '2019-03-06T12:49:58.781Z',
-    'offerers?page=2': '2019-03-06T09:50:37.735Z',
-    'offerers?page=3': '2019-03-06T09:50:41.369Z',
-    'offerers?page=4': '2019-03-06T09:50:41.666Z',
-    'offers/AUSQ': '2019-03-06T16:08:06.116Z',
-    offerers: '2019-03-06T16:32:01.204Z',
-    'offers/AWHA': '2019-03-06T17:14:53.837Z',
-    'offers/UU': '2019-03-07T13:02:50.963Z',
-  },
+  tracker: {},
   user: {
     currentUserUUID: '22a18e44-710c-45cd-b442-34d8ff417fac',
     id: 'FE',
@@ -1288,7 +1079,7 @@ const mockedState = {
     isAdmin: true,
     lastName: '93 0',
     modelName: 'User',
-    phoneNumber: '0612345678',
+    phoneNumber: '67856557574',
     postalCode: '93100',
     publicName: 'PC Test Admin 93 0',
     thumbCount: 0,
