+++ src/components/pages/Offer/RawEventOccurrencesAndStocksManager.jsx
@@ -8,8 +8,14 @@ import PropTypes from 'prop-types'
 import EventOccurrenceAndStockItem from './EventOccurrenceAndStockItem'
 import HeroSection from '../../layout/HeroSection'
 
-export function getAddUrl(isEditing, isStockOnly, offerId, stocks, defaultUrl) {
-  if (isEditing) {
+export function getAddUrl(
+  editedStockId,
+  isStockOnly,
+  offerId,
+  stocks,
+  defaultUrl
+) {
+  if (editedStockId) {
     return defaultUrl
   }
 
@@ -39,7 +45,7 @@ class RawEventOccurrencesAndStocksManager extends Component {
   }
 
   handleEscKey() {
-    if (!this.props.isEditing) {
+    if (!this.props.editedStockId) {
       this.onCloseClick()
     } else {
       const cancelButton = document.getElementsByClassName('cancel-step')[0]
@@ -49,13 +55,13 @@ class RawEventOccurrencesAndStocksManager extends Component {
 
   handleEnterKey() {
     // Dirty DOM selectors but...
-    if (!this.props.isEditing) {
+    if (!this.props.editedStockId) {
       // This dom node is generated by react-router-dom's <NavLink /> component wich is
       // a 'Function component' and...
       // Warning: Function components cannot be given refs. Attempts to access this ref will fail.
       document.getElementById('add-occurrence-or-stock').focus()
-      const { isEditing, isStockOnly } = this.props
-      this.props.history.push(this.getAddUrl(isEditing, isStockOnly))
+      const { editedStockId, isStockOnly } = this.props
+      this.props.history.push(this.getAddUrl(editedStockId, isStockOnly))
     } else {
       // Could fetch ref from the included components, but it will be far more complex
       // as we would have to make the ref transit through callback props.
@@ -66,9 +72,9 @@ class RawEventOccurrencesAndStocksManager extends Component {
     }
   }
 
-  getAddUrl(isEditing, isStockOnly) {
+  getAddUrl(editedStockId, isStockOnly) {
     return getAddUrl(
-      isEditing,
+      editedStockId,
       isStockOnly,
       get(this.props.offer, 'id'),
       this.props.stocks,
@@ -103,7 +109,7 @@ class RawEventOccurrencesAndStocksManager extends Component {
       errors,
       event,
       eventOccurrences,
-      isEditing,
+      editedStockId,
       isNew,
       provider,
       stocks,
@@ -191,9 +197,9 @@ class RawEventOccurrencesAndStocksManager extends Component {
                   <td colSpan="10">
                     <NavLink
                       className="button is-secondary"
-                      disabled={isEditing}
+                      disabled={editedStockId}
                       id="add-occurrence-or-stock"
-                      to={this.getAddUrl(isEditing, isStockOnly)}>
+                      to={this.getAddUrl(editedStockId, isStockOnly)}>
                       {isStockOnly
                         ? stocks.length
                           ? ''
@@ -264,7 +270,6 @@ RawEventOccurrencesAndStocksManager.defaultProps = {
 }
 
 RawEventOccurrencesAndStocksManager.propTypes = {
-  isEditing: PropTypes.bool.isRequired,
   stocks: PropTypes.array,
   eventOccurrences: PropTypes.array,
 }
